requirements.txt
requests
openai
langchain
typing-extensions
langgraph
telebot
python-logging-loki
python-dotenv
influxdb-client
langchain-groq
langchain-openai